backend/requirements.txt
flask==2.3.3
mysql-connector-python==8.0.33
